+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/ConfigImpl.java
@@ -554,8 +554,10 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class ControllerEntryImpl implements ControllerEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.CONTROLLER_CAPACITY;
 
+        @Override
         public long getEnergyCapacity() {
             return energyCapacity;
         }
@@ -708,6 +710,7 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class WirelessGridEntryImpl implements WirelessGridEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.WIRELESS_GRID_CAPACITY;
 
         private long openEnergyUsage = DefaultEnergyUsage.WIRELESS_GRID_OPEN;
@@ -754,6 +757,7 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class PortableGridEntryImpl implements PortableGridEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.PORTABLE_GRID_CAPACITY;
 
         private long openEnergyUsage = DefaultEnergyUsage.PORTABLE_GRID_OPEN;
@@ -855,6 +859,7 @@ public class ConfigImpl implements ConfigData, com.refinedmods.refinedstorage.co
     }
 
     private static class WirelessAutocraftingMonitorEntryImpl implements WirelessAutocraftingMonitorEntry {
+        @ConfigEntry.BoundedDiscrete(min = 1L, max = Long.MAX_VALUE)
         private long energyCapacity = DefaultEnergyUsage.WIRELESS_AUTOCRAFTING_MONITOR_CAPACITY;
 
         private long openEnergyUsage = DefaultEnergyUsage.WIRELESS_AUTOCRAFTING_MONITOR_OPEN;